files for working with stacked tif images
folder includes example fish images and associated code for binarizing
images and reading in and processing stacked binary tif files.

- binarizeImage.R takes a simple .png file and binarizes it
- tif2coords.R reads in both single and stacked tif files and extracts
coordinates for the outline of the image. Also plots in 3d for a quick
look
-testImages has the original fish image, the non-binarized .tif file,
and the binarized .tif file
-stackedTIF mimicks the structure of the coupon surface data. This
folder is read into 'tif2coords.R' to create a 3d fish