update histogram criteria for finding layers
iterates over a number of z axis rotations to find layers in a 45 degree
coupon, both synthetic and actual. Plots the number of zero bins for
each transformed z axis histogram so the optimal turning angle can be
found. Also contains code for plotting the histogram associated with the
optimal turning angle.